refactor: migrated ColorPicker to svelte-undp-components
