+++ .changeset/moody-masks-begin.md
@@ -1,0 +1,6 @@
+---
+"@undp-data/svelte-undp-components": patch
+"geohub": patch
+---
+
+refactor: migrated ColorPicker to svelte-undp-components
