+++ packages/svelte-undp-components/package.json
@@ -96,6 +96,7 @@
 		"esm-env": "^1.0.0",
 		"lodash-es": "^4.17.21",
 		"maplibre-gl": "^4.7.1",
+		"svelte-awesome-color-picker": "^3.1.4",
 		"svelte-copy": "^1.4.2",
 		"svelte-range-slider-pips": "^3.1.1",
 		"svelte-tippy": "^1.3.2",
